move node:net implementation to internal folder
